Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.5",
    "alembic>=1.18.4",
    "anthropic>=0.96.0",
    "asyncio>=4.0.0",
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.14.3",
    "celery>=5.6.3",
    "duckduckgo-search>=8.1.1",
    "fake-useragent>=2.2.0",
    "fastapi>=0.136.0",
    "flower>=2.0.1",
    "google-api-python-client>=2.194.0",
    "groq>=1.2.0",
    "html5lib>=1.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "linkedin-api>=2.3.1",
    "loguru>=0.7.3",
    "lxml>=5.4.0",
    "nltk>=3.9.4",
    "numpy>=1.26.3",
    "pandas>=2.3.3",
    "passlib[bcrypt]>=1.7.4",
    "pdfplumber>=0.11.9",
    "pillow>=12.2.0",
    "playwright>=1.58.0",
    "psycopg2-binary>=2.9.12",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "pypdf2>=3.0.1",
    "python-dateutil>=2.9.0.post0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-jobspy>=1.1.82",
    "python-jose[cryptography]>=3.5.0",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.26",
    "redis>=7.4.0",
    "requests>=2.33.1",
    "resend>=2.29.0",
    "scikit-learn>=1.8.0",
    "scrapy>=2.15.0",
    "sentence-transformers>=5.4.1",
    "sentry-sdk>=2.58.0",
    "spacy>=3.8.14",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.45.0",
]
